app/requirements.txt
pymilvus==2.3.6
sentence-transformers==3.0.1
torch>=2.1
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
marshmallow>=3.13,<4
environs
numpy<2

